# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Duplicate Payment Detection System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /training/sessions:
    get:
      operationId: getTrainingSessions
      tags: [training]
      summary: Get all training sessions
      responses:
        "200":
          description: List of training sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSessionListResponse"
    post:
      operationId: createTrainingSession
      tags: [training]
      summary: Create a new training session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingSessionBody"
      responses:
        "200":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSession"
  /training/sessions/{id}/messages:
    post:
      operationId: sendTrainingMessage
      tags: [training]
      summary: Send a message in a training session to teach the agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingMessageBody"
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingMessageResponse"
...
components:
  schemas:
...
    TrainingSession:
      type: object
      properties:
        id:
          type: string
        trainingType:
          type: string
          enum: [database_schema, duplicate_definition, general]
        title:
          type: string
        createdAt:
          type: string
        messageCount:
          type: integer
        lastMessageAt:
          type: string
      required: [id, trainingType, title, createdAt, messageCount]
    TrainingSessionListResponse:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/TrainingSession"
      required: [sessions]
    CreateTrainingSessionBody:
      type: object
      properties:
        trainingType:
          type: string
          enum: [database_schema, duplicate_definition, general]
        title:
          type: string
      required: [trainingType, title]
...
    TrainingMessageBody:
      type: object
      properties:
        message:
          type: string
      required: [message]
    TrainingMessageResponse:
      type: object
      properties:
        response:
          type: string
        memorySaved:
          type: boolean
        memoryKey:
          type: string
      required: [response, memorySaved]
...